fix all the things

+++ src/basic_recipes/ablines.jl
@@ -36,8 +36,8 @@ function Makie.plot!(p::ABLines)
     p
 end
 
-data_limits(::ABLines) = Rect3f(Point3f(NaN), Vec3f(NaN))
-boundingbox(::ABLines, space::Symbol = :data) = Rect3f(Point3f(NaN), Vec3f(NaN))
+data_limits(::ABLines) = Rect3d(Point3f(NaN), Vec3f(NaN))
+boundingbox(::ABLines, space::Symbol = :data) = Rect3d(Point3f(NaN), Vec3f(NaN))
 
 function abline!(args...; kwargs...)
     Base.depwarn("abline! is deprecated and will be removed in the future. Use ablines / ablines! instead." , :abline!, force = true)
